feat: overhaul ordering UI and fix critical system-level print blocking • feat: implement a 100% custom in-app overlay for confirmation dialogs to bypass Android 10 WRITE_SETTINGS security errors • feat: support category-based menu organization with a tabbed filtering UI for high-volume menu management • feat: upgrade importMenus logic to handle structured JSON with category grouping and robust price parsing  refactor: enhance menu visibility using a 2-column adaptive grid with larger fonts and explicit allergy warnings  • fix: stabilize Bluetooth communication using address-based device matching and dual RFCOMM connection attempts (Secure/Insecure)  • fix: resolve price calculation and string formatting errors introduced by switching to Double currency types  • style: apply professional OrderFlow branding and remove redundant legacy XML icon assets to ensure clean deployment

+++ app/src/main/java/com/remotemenu/model/MenuItem.kt
@@ -5,14 +5,14 @@ package com.remotemenu.model
  * 매장에서 판매하는 메뉴의 정보를 담는 데이터 클래스.
  * @property id 메뉴 고유 식별자
  * @property name 메뉴 이름
- * @property price 메뉴 가격
+ * @property price 메뉴 가격 (소수점 지원)
  * @property allergy 포함된 알러지 정보 문자열
  * @property customOptions 선택 가능한 추가 옵션 리스트
  */
 data class MenuItem(
     val id: Int,
     val name: String,
-    val price: Int,
+    val price: Double,
     val allergy: String,
     val customOptions: List<CustomOption>
 )
